Add capability to exclude reports by setting a flag in the base class. Update SimpleDF to use it, and the SimpleDF example input files to include an extra column to represent it.
We're now ready to start creating a GUI program to use this library.

The Doxygen comments for computeLeastSquaresFix need to be updated to
include the math involved.  The function evaluates the 2x2 matrix in a
way that is not at all obvious, and which represents the final step in a
long line of matrix-matrix multiplies.