Add a new derived class off of DF_Report_Collection.
DF_Report_Collection takes DFLib::Abstract::Report pointers because it simply
doesn't care what type of report you give it as long as it has the basic
interface it needs.  But this makes certain other operations on the pointers
it holds difficult for applications unless they maintain copies of those
pointers in some other container.  That is a prescription for ugliness and
angry Gods.

DFLib::ProjReportCollection is a DF_Report_Collection that insists on being
given DFLib::Proj::Report pointers.  Under the hood it simply passes them
down to the base class methods.  But now we can safely assure ourselves that
every item in the collection is of the right type, and can always call
certain methods on them.

Also, make some important methods of the base class virtual, so we can
overload them properly.