comment out gridfile writing.
Looks like the fixes to DF_ReportCollection that make sure the bearing error
is in a correct range were exactly the source of bad fixes.  The ML fix for
the SAR Support ELT practice is now almost right on top of the actual location
of the transmitter, whereas before the fix it was almost a mile away.

Plotting the cost function as a surface was key to seeing this problem, so
leaving the code in.