Fix dumb-ass bug in Nelder-Mead (computation of centroid).
Still no joy in finding a generic starting guess for conjugate gradients.

Probably the best bet here is to just punt and ignore ML fixes that are
more than a certain distance from any receiver, say 100 miles?  Dunno.

There's *gotta* be a way to force this to converge to sane numbers.

The problem is that in some pathological cases the Brent minimization takes
us way out to infinity because the function is so flat.  Perhaps there's the
key.  Where's the lock?