Time to punt the whole convergence thing.  You will get lost here.
Slam in a quick and dirty computation of the distance between the ML fix
and one selected receiver.  If it's more than 100 miles, flag as invalid and
just move on.

Do something similar in qDF, only make it so that the ML fix shows in the
display as invalid and any object that's been created is deleted --- no
APRS display of crap fixes!