Re-enable Nelder-Mead pre-minimization before trying conjugate gradients.
This does not fix all the convergence problems I've seen, but it helps the ones
that actually have a minimum to find.  I am now convinced that when this
combined effort fails, there simply is no real minimum and it's a waste of
energy to try to find one.

The only thing to be done is to reject "fixes" that are absurd and just flag
them as invalid so they're not displayed.