Remove all includes of "projects.h" and replace with "proj_api.h"
Add a declaration of the non-api call "dmstor" from libproj, since that was
the only thing we needed projects.h for.

Remove "#undef XY" since that was only there because projects.h polluted the
macro namespace with such a tiny thing, and it screwed up all our uses of
"XY" as a suffix.