Make "testlsDF_proj" output the ML confidence intervals as well as the Stansfield.
Try running
     testlsDF_proj "106d35'W" "34d55'40\"N" < receivers6
to see this stuff in action.

You can plot the results with
  gnuplot testlsDFfix.gnuplot

For most of the run with this receiver set, the Stansfield and ML
fixes are pretty close, as are the ellipses.  To get more dramatic
differences between the two, you have to have large bearing error or
short distances (Stansfield assumes small bearing error and large
distances).