Very, very early cut at SWIG Python bindings for DFLib.
Only the abstract point, abstract report, and report collection are
wrapped.  I have NOT yet tested the functionality of the
collection.

The intent here is provide a barebones Python library that can be used
to create Python implementations of the point and report classes, and
use them as-is in the report collection.

I *think* the classes in pyPointTest.py will work in the same way that
the DFLib::XY point and report classes do, but have not had time to
test them yet.  But the current set of wrapped classes seems to link
and load into Python just fine, and the Python classes derived from
DFLib abstract interfaces seem to work.