Modify DFLib.i so it prints more useful output for exceptions (I had a lot of errors due to not getting the "Clone" function written correctly, and all I got for debugging without this change was a notice that an exception happened).
Add actual DFLib fix computation tests to "pyPointTest.py".

These tests require the "pyproj" package, http://jswhit.github.io/pyproj/

The test performed by pyPointTest is basically the same as a run of
testlsDFfix with the transmitter at 106d35'W 34d55'N, and one particular
randomization of bearings.  Confirms that this python code is producing
the same fixes as the C++ code does.

The python is a little fragile, but it does work, and could probably
be used as is.  One could refine the definitions of the pyPoint and
pyReport classes to do the same sort of "user coordinate <-> XY"
coordinate stuff that the fancier DFLib classes do.  Or just use them
as is and do the coordinate conversion yourself.  The main advantage
of doing the coordinate conversion in the "Point" class would be to
make sure the FCA standard deviations are computed in user coordinate
units rather than XY units, as the deviations are pretty useless in XY
coordinates.