Add a Windows installer of the DFLib python library.
With this installer, Losedows users will be able to install the python
bindings for DFLib directly into their existing python install without
having to go through the trouble of downloading the outdated version of
Visual Studio needed to build it.

Simply run the MSI installer in the "Python/dist" directory, and you will
be able to import DFLib in python scripts under Windows.

Not going to do this for users of real operating systems, though.
Building DFLib's python bindings on other systems should be much more
straightforward.