Make configure search for non-gcc compilers
FreeBSD uses clang by default.

When building qDF with FreeBSD's qmake defaults, qDF gets built with
clang++.  When building DFLib with configure's ddefaults, DFLib was
getting built with g++.  The resulting mismatch in shared libraries
required confused everything.

Now, configure will search for compilers in this order:
  c++ CC g++ clang++

and C compilers:
  cc gcc clang

If some other compiler is desired, just specify it on the configure line
with CC= and CXX=.

Since FreeBSD is my primary platform and I seem to be the only person
using this, I'm fine with making it work with no command line options
for me.  So.  There.