Update DFLib to use modern PROJ library
This commit changes all use of PROJ to be compatible with the new
Proj.6 API.  This breaks its compatibility with the old Proj.4 API,
but who cares?

Add the ProjUnitTests program to the make but doesn't install.

Updates all test programs to work properly.

All of the "testlsDFfix_*" programs run and produce the same output
when the random number generator is given the same seed.

For example:

./testlsDF_proj --seed 1767062539 "106d25'W" "34d57'N" < ../receivers6_sloppy

and

./testlsDF_ll --seed 1767062539 "106d25'W" "34d57'N" < ../receivers6_sloppy

give the same locations for all fixes, though the _proj one produces
more output and does more stuff (including creation of confidence
ellipses in the .gnuplot file)

SimpleDF and SimpleDF2 produce correct output when given the
"ELTPractice" input file (actual data from an ELT practice session
from ages ago).

Closes #1