Remove gdal use, tweak pyProjPoint.py
testlsDF_proj.cpp once upon a time used GDAL to create a GeoTIFF file
of the "cost function" used to compute the Maximum Likelihood Fix as
if it were an elevation map.

That was only for my debugging, to try to view this in a GIS to see if
I could examine the landscape of the cost function to understand
convergence problems.  It never was very enlightening and I eventually
commented it all out because in at least one version of gdal I was
getting valgrind errors from what I was doing.

I'm removing all that junk, as it just bloats the test code and
configure.

Added a "--norandom" command line option to testlsDF_proj so one can
run the thing without randomizing bearings --- that is, it will
compute the bearings from the receivers to the target exactly, and
plug those right in to the DF report.  You should wind up with fixes
that pretty much are smack dab on the transmitter.

Also, got pyProjPoint working and printing the ML fix in DD MM.MMM H
format.

Finally, testlsDF_proj was also outputting a "grasspoints" file, which
was a cheesy vector format I used to pull into grass6 ages ago for
visualization of the data.  I don't use grass for anything anymore,
and grass6 is ancient.  That stuff is gone.